refactor: separated validation helper
