+++ src/test/property.test.ts
@@ -4,14 +4,13 @@ import { mock, MockProxy } from 'jest-mock-extended';
 import { App } from '@/app';
 import { PropertyRoute } from '@routes/property.route';
 import { PropertyService } from '@/services/property.service';
-import { getPropertyTypes, getPropertyPurpose, isInvalidNumber } from '@/utils/helpers';
+import { getPropertyTypes, getPropertyPurpose } from '@/utils/helpers';
 import { AVAILABLE_CITIES } from '@/types';
 
 jest.mock('@/services/property.service');
 jest.mock('@/utils/helpers', () => ({
   getPropertyTypes: jest.fn(),
   getPropertyPurpose: jest.fn(),
-  isInvalidNumber: jest.fn(),
 }));
 const propertyServiceMock: MockProxy<PropertyService> = mock<PropertyService>();
 
@@ -90,7 +89,6 @@ describe('Property', () => {
   });
   describe('GET /property', () => {
     it('should retrieve properties with default pagination and sorting', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockfindAllProperties();
       const response = await request(app).get('/property').query({ purpose: 'for_sale' });
       expect(response.status).toBe(200);
@@ -103,7 +101,6 @@ describe('Property', () => {
     });
 
     it('should retrieve properties with custom pagination', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockfindAllProperties();
       const response = await request(app).get('/property?page_size=5&page_number=2').query({ purpose: 'for_sale' });
       expect(response.status).toBe(200);
@@ -114,7 +111,6 @@ describe('Property', () => {
     });
 
     it('should retrieve properties sorted by price in descending order', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockfindAllProperties();
       const response = await request(app).get('/property?sort_by=price&sort_order=DESC').query({ purpose: 'for_sale' });
       expect(response.status).toBe(200);
@@ -126,19 +122,16 @@ describe('Property', () => {
       expect(prices).toEqual([...prices].sort((a, b) => b - a));
     });
     it('should return 400 for invalid pagination parameters', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(true);
       const response = await request(app).get('/property?page_size=-1&page_number=0');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
     });
     it('should return 400 for invalid sorting parameters', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       const response = await request(app).get('/property?sort_by=unknown_field&sort_order=DESC');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
     });
     it('should return 400 for invalid sorting parameters (sort_order)', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       const response = await request(app).get('/property?sort_by=price&sort_order=unknown');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
@@ -152,7 +145,6 @@ describe('Property', () => {
   });
   describe('GET /property/count', () => {
     it('should retrieve the total count of properties', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       propertyServiceMock.getPropertiesCountMap.mockResolvedValue({
         'Agricultural Land': 76,
         Building: 2618,
@@ -166,7 +158,6 @@ describe('Property', () => {
       expect(response.body).toHaveProperty('message');
     });
     it('should retrieve the total count of properties', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       propertyServiceMock.getPropertiesCountMap.mockResolvedValue({
         'Agricultural Land': 76,
         Building: 2618,
@@ -180,7 +171,6 @@ describe('Property', () => {
       expect(response.body).toHaveProperty('message');
     });
     it('Should return error', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       propertyServiceMock.getPropertiesCountMap.mockRejectedValue('Error');
       const response = await request(app).get('/property/count').query({ purpose: 'for_sale' });
       expect(response.status).toBe(500);
@@ -269,7 +259,6 @@ describe('Property', () => {
   });
   describe('GET /property/search', () => {
     it('should retrieve properties with default pagination and sorting', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockSearchProperties();
       const response = await request(app).get('/property/search?purpose=for_sale');
       expect(response.status).toBe(200);
@@ -282,7 +271,6 @@ describe('Property', () => {
     });
 
     it('should retrieve properties with custom pagination', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockSearchProperties();
       const response = await request(app).get('/property/search?page_size=5&page_number=2&purpose=for_sale');
       expect(response.status).toBe(200);
@@ -293,7 +281,6 @@ describe('Property', () => {
     });
 
     it('should retrieve properties sorted by price in descending order', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockfindAllProperties();
       const response = await request(app).get('/property/search?sort_by=price&sort_order=DESC&purpose=for_sale');
       expect(response.status).toBe(200);
@@ -305,7 +292,6 @@ describe('Property', () => {
       expect(prices).toEqual([...prices].sort((a, b) => b - a));
     });
     it('should return 400 for invalid pagination parameters', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(true);
       const response = await request(app).get('/property/search?page_size=-1&page_number=0');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
@@ -321,7 +307,6 @@ describe('Property', () => {
       expect(response.body).toHaveProperty('message');
     });
     it('should search', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockSearchProperties();
       const responses = await Promise.all([
         request(app).get('/property/search?query=islamabad&purpose=for_sale'),
@@ -336,31 +321,26 @@ describe('Property', () => {
       });
     });
     it('should return 400 for invalid price_min parameter', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(true);
       const response = await request(app).get('/property/search?price_min=-1');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
     });
     it('should return 400 for invalid price_max parameter', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(true);
       const response = await request(app).get('/property/search?price_max=-1');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
     });
     it('should return 400 for invalid area_min parameter', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(true);
       const response = await request(app).get('/property/search?area_min=-1');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
     });
     it('should return 400 for invalid area_max parameter', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(true);
       const response = await request(app).get('/property/search?area_max=-1');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
     });
     it('should return 400 for invalid bedrooms parameter', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(true);
       const response = await request(app).get('/property/search?bedrooms=1,a');
       expect(response.status).toBe(400);
       expect(response.body).toHaveProperty('message');
@@ -381,7 +361,6 @@ describe('Property', () => {
       expect(response.body).toHaveProperty('message');
     });
     it('Should return error', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       propertyServiceMock.searchProperties.mockRejectedValue('Error');
       const response = await request(app).get('/property/search?query=islamabad').query({ purpose: 'for_sale' });
       expect(response.status).toBe(500);
@@ -405,7 +384,6 @@ describe('Property', () => {
   });
   describe('GET /property/featured', () => {
     it('Should return featured properties', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       mockSearchProperties();
       const response = await request(app).get('/property/featured?purpose=for_sale');
       expect(response.status).toBe(200);
@@ -415,7 +393,6 @@ describe('Property', () => {
       expect(response.body.data).toHaveProperty('total_count');
     });
     it('Should return error', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       propertyServiceMock.searchProperties.mockRejectedValue('Error');
       const response = await request(app).get('/property/featured').query({ purpose: 'for_sale' });
       expect(response.status).toBe(500);
@@ -424,7 +401,6 @@ describe('Property', () => {
   });
   describe('GET /property/similar', () => {
     it('Should return featured properties', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       propertyServiceMock.findPropertyById.mockResolvedValue([
         {
           id: 1,
@@ -447,7 +423,6 @@ describe('Property', () => {
       expect(response.body.data).toHaveProperty('total_count');
     });
     it('Should return error', async () => {
-      (isInvalidNumber as jest.Mock).mockReturnValue(false);
       propertyServiceMock.searchProperties.mockRejectedValue(new Error('This should not be done!'));
       const response = await request(app).get('/property/similar?id=1').query({ purpose: 'for_sale' });
       expect(response.status).toBe(500);
