+++ src/utils/validation.helpers.ts
@@ -1,0 +1,1 @@
+export const isInvalidNumber = (value: string): boolean => isNaN(Number(value)) || Number(value) < 0;
